+++ next/web/src/App/Admin/Tickets/Ticket/Timeline/OpsLog.tsx
@@ -21,7 +21,7 @@ import { Button, Spin, Tag } from '@/components/antd';
 import { UserLabel } from '@/App/Admin/components';
 import { useModal } from '../components/useModal';
 import { useTicketFields_v1 } from '../api1';
-import { Time } from './Time';
+import { Time } from '../components/Time';
 
 interface OpsLogProps<Action extends OpsLogSchema['action'] = any> {
   data: Extract<OpsLogSchema, { action: Action }>;
@@ -246,7 +246,7 @@ function DiffFields({ changes }: DiffFieldsProps) {
     );
   }
 
-  return (changes.map(({ fieldId, from, to }) => {
+  return changes.map(({ fieldId, from, to }) => {
     const field = fieldById[fieldId];
     if (!field) {
       return null;
@@ -261,7 +261,7 @@ function DiffFields({ changes }: DiffFieldsProps) {
         to={to}
       />
     );
-  }) as any) as JSX.Element;
+  }) as any as JSX.Element;
 }
 
 type OptionValue = string;
