+++ next/web/src/App/Admin/Tickets/Ticket/components/ReplyCard.tsx
@@ -5,7 +5,7 @@ import { Image } from 'antd';
 import { partition } from 'lodash-es';
 import cx from 'classnames';
 
-import { Time } from '../Timeline/Time';
+import { Time } from './Time';
 
 const IMAGE_FILE_MIMES = ['image/png', 'image/jpeg', 'image/gif'];
